fix: always calculate child rel field offsets for join predicates
The confusion for me has been the location of the `field` lookup,
referenced in the
`expression->scalar_function->arguments->value->selection->direct_reference->struct_field`

Where is the list of columns that pokes in to?

It doesn't explicitly exist -- it is the flattened list of all columns
from the
left and the right tables.

We were correctly calculating the `relative_offset` -- the offset within the
left- or right-table -- but our `base_offset` was wrong.  Why?

Because in some situations, we don't have the `child_rel_field_offsets`
to pass to the `translate` call for the _join predicates_ and they need
to be there.

So we check if that has been calculated, and if it hasn't, then we
calculate them and pass them in.