feat: add initial support for ibis 4.x
A lot of this is semi-gross hacks, but they should be relatively easy to
unwind when we drop support for Ibis 3.x (whenever that happens).

Changes:

The biggest hurdle is that we no longer have the expr -> op -> expr ...
pattern, it's just ops all the way down.

To maintain compatibility, any reference to something like `op.arg` is
now written as `op.arg.op().to_expr()` -- this is gross, but functional.
It's a very little bit of extra work, but it makes sure that we get the
same object no matter the Ibis version.

Also, because now we're often passing around ops instead of exprs, or
because the expr isn't strictly necessary for a bunch of dispatch rules,
both `expr` and `compiler` are both keyword arguments now.

Besides that there are some minor shenigans to handle the nonexistence
of `ops.CountStar` in ibis 3.x, and the renaming / rearrangement of the
graph utilities.

All in all, there are only a few explicit 4.x vs 3.x branches.