chore(ci): separate ibis jobs, only run on lowest and highest supported
Separates the Ibis 3.x and Ibis 4.x into separate jobs, because we can't
support the same matrix of Python versions in both.

For both 3.x and 4.x, we want to run on the lowest and highest support
versions of Python.

For the `nix` job, we run on all supported versions with ibis 4 to generate the
files for cachix.