refactor: remove expr arguments in translate
This removes all usage of `expr` being passed explicitly to mapping
functions.  There are a few cases where Ibis 3.x will pass in
expressions still, but those should be handled by the `_expr` rule,
which forwards on the `op` (at a cost of an extra dispatch).

There are also a few bits where we do reductions on predicates that
operate on expressions.  Those should be converted to ops where
possible (without breaking support for older Ibis versions).