feat: add type signature to scalar function name definitions
As per a recent change in the spec upstream (and also to enable better
compatibility with other substrait libraries), extension function names
now have the expected input types included in the name of the function.

This can seem very weird, but helps to make it much easier for consumers
to parse a plan.

Basically, if you call a scalar function `add` and pass in two integer
32s, the "name" of the function will be `add:i32_i32`.

We do not validate that the particular signature exists in the
extension.yaml files, but we do perform the requisite upcasting to make
sure that only integers and floats of the same precision are passed to
functions (also in line with the Substrait specification).