fix(extensions): use compound signature when creating new URI
This fixes a bug that I missed when I reintroduced the compound function
signatures.

If we call `add` using int32 and float32, respectively, then we should
have two separate scalar function anchors, one that is `add:i32_i32` and
another that is `add:fp32_fp32`.  This is now handled correctly.

There is a bit of uncertainty about how we handle the signatures for
variadic functions, especially ones that accept (optionally) zero
arguments.

What I have here is my best guess as to the right way to do it, and I
can update it after consulting with the Substrait community at the next meeting.